+++ src/lib/components/contacts/GeneralInfo.svelte
