+++ src/routes/events/+page.svelte
@@ -1,8 +1,6 @@
 <script>
-    import events, { Tag } from '$lib/data/events';
-    import FeaturedEvent from '$lib/components/headers/FeaturedEvent.svelte';
+    import events from '$lib/data/events';
     import FilterPagePanel from '$lib/components/panels/FilterPagePanel.svelte';
-    import lino from '$lib/assets/lino/lino-sablay.svg';
 
     const filters = [
         'All',
@@ -15,17 +13,6 @@
     ];
 </script>
 
-<article class="prose flex max-w-full flex-col gap-8">
-    <FeaturedEvent tags={[Tag.App]} schedule={['']}>
-        {#snippet image()}
-            <img src={lino} alt="Incoming App Process" />
-        {/snippet}
-        {#snippet name()}
-            Incoming App Process...
-        {/snippet}
-        {#snippet description()}
-            Stay Tuned...
-        {/snippet}
-    </FeaturedEvent>
+<section class="prose flex max-w-full">
     <FilterPagePanel {filters} cardsInfo={events} />
-</article>
+</section>
