+++ src/routes/projects/+page.svelte
@@ -1,9 +1,9 @@
 <script>
-    import Faq from '$lib/components/panels/Faq.svelte';
-    import FaqEntry from '$lib/components/items/FaqEntry.svelte';
+    import FaqItem from '$lib/components/items/FaqItem.svelte';
+    import FaqPanel from '$lib/components/panels/FaqPanel.svelte';
     import FilterPagePanel from '$lib/components/panels/FilterPagePanel.svelte';
-    import GeneralInfo from '../../lib/components/miscellaneous/GeneralInfo.svelte';
-    import SocialMediaPanel from '$lib/components/panels/SocialMediaPanel.svelte';
+    import GeneralInfo from '../../lib/components/contacts/GeneralInfo.svelte';
+    import SocialMedia from '$lib/components/contacts/SocialMedia.svelte';
     import projects from '$lib/data/projects';
 
     const filters = ['All', 'Service', 'Innovation', 'Other'];
@@ -15,23 +15,23 @@
     </div>
     <div class="flex flex-col gap-4">
         <GeneralInfo />
-        <SocialMediaPanel />
+        <SocialMedia />
     </div>
 </section>
 
 <section>
-    <Faq>
-        <FaqEntry title="What kind of projects can I commission?">
+    <FaqPanel>
+        <FaqItem title="What kind of projects can I commission?">
             <span>UP CSI is open to accepting web development projects as of the moment. </span>
-        </FaqEntry>
-        <FaqEntry title="How much do projects usually cost?">
+        </FaqItem>
+        <FaqItem title="How much do projects usually cost?">
             <span
                 >Costs vary depending on the scale of the project and the customer’s budget. Prices
                 are usually negotiable as projects made with UP CSI is also a way for its members to
                 practice the skills they have learned.</span
             >
-        </FaqEntry>
-        <FaqEntry title="How do I commission a project?">
+        </FaqItem>
+        <FaqItem title="How do I commission a project?">
             <span
                 >You may contact UP CSI by sending an email <a href="mailto:lino@up-csi.org"
                     >here.</a
@@ -45,8 +45,8 @@
                 <li>Other project-specific concerns</li>
             </ul>
             <span> UP CSI will get back to you to set an appointment within 5 business days. </span>
-        </FaqEntry>
-        <FaqEntry title="Who will work on my project?">
+        </FaqItem>
+        <FaqItem title="Who will work on my project?">
             <span
                 >The Director of Engineering and the Vice President for Service decide who will work
                 on the projects. Priority is given to members who haven’t participated in a service
@@ -55,6 +55,6 @@
                 members, who are mostly passionate students from the UP Diliman Department of
                 Computer Science, make up the pool of candidates for these projects.</span
             >
-        </FaqEntry>
-    </Faq>
+        </FaqItem>
+    </FaqPanel>
 </section>
